create pattern dir for isolating codes
